feat: edit modal in myboard page -ing
